[LLVMExternalProjects] Drop the IREE LinalgTransformDialect
The LinalgTransformDialect has become a maintenance burden without any
users. This drops the dialect and cleans up some surrounding linalg ext
code that has since been upstreamed in some capacity.

 - Transform ops equivalent to lower_to_llvm now exist upstream
 - The linalg ext variant of the TransformDialectInterpreter pass has
   since been moved to IREE core

take_first, emit_remark, and the matcher infra + ops are still load
bearing and transitioning to upstream is left as TODO.
