+++ compiler/src/iree/compiler/API/api_exports.def
@@ -339,7 +339,6 @@ EXPORTS
   mlirFunctionTypeGetTypeID
   mlirGetDialectHandle__iree_input__
   mlirGetDialectHandle__iree_linalg_ext__
-  mlirGetDialectHandle__iree_linalg_transform__
   mlirGetDialectHandle__transform__
   mlirIREELinalgTransformRegisterPasses
   mlirIREETransformRegisterPasses
