+++ compiler/src/iree/compiler/API/api_exports.ld
@@ -340,7 +340,6 @@ VER_0 {
     mlirFunctionTypeGetTypeID;
     mlirGetDialectHandle__iree_input__;
     mlirGetDialectHandle__iree_linalg_ext__;
-    mlirGetDialectHandle__iree_linalg_transform__;
     mlirGetDialectHandle__transform__;
     mlirIREELinalgTransformRegisterPasses;
     mlirIREETransformRegisterPasses;
