+++ compiler/src/iree/compiler/API/api_exports.macos.lst
@@ -338,7 +338,6 @@ _mlirFunctionTypeGetResult
 _mlirFunctionTypeGetTypeID
 _mlirGetDialectHandle__iree_input__
 _mlirGetDialectHandle__iree_linalg_ext__
-_mlirGetDialectHandle__iree_linalg_transform__
 _mlirGetDialectHandle__transform__
 _mlirIREELinalgTransformRegisterPasses
 _mlirIREETransformRegisterPasses
