+++ llvm-external-projects/iree-dialects/include/iree-dialects-c/Dialects.h
@@ -31,8 +31,6 @@ MLIR_DECLARE_CAPI_DIALECT_REGISTRATION(IREELinalgExt, iree_linalg_ext);
 // LinalgTransform
 //===--------------------------------------------------------------------===//
 
-MLIR_DECLARE_CAPI_DIALECT_REGISTRATION(LinalgTransform, iree_linalg_transform);
-
 /// Register all passes for LinalgTransform.
 MLIR_CAPI_EXPORTED void mlirIREELinalgTransformRegisterPasses();
