+++ llvm-external-projects/iree-dialects/include/iree-dialects/Dialect/LinalgTransform/StructuredTransformOpsExt.td
@@ -15,36 +15,6 @@ include "mlir/Interfaces/SideEffectInterfaces.td"
 include "mlir/IR/OpAsmInterface.td"
 include "mlir/IR/OpBase.td"
 
-def LowerToLLVMOp : Op<Transform_Dialect, "lower_to_llvm",
-    [FunctionalStyleTransformOpTrait,
-     MemoryEffectsOpInterface,
-     DeclareOpInterfaceMethods<TransformOpInterface>,
-     ReportTrackingListenerFailuresOpTrait]> {
-  let description = [{
-    Indicates that the entire targeted module should be converted
-    to the LLVM dialect. This is expected to be the last transformation in
-    a sequence.
-
-    #### Return modes
-
-    This operation consumes the `target` handle and produces the `transformed`
-    handle.
-  }];
-
-  let arguments =
-    (ins TransformHandleTypeInterface:$target,
-     DefaultValuedAttr<BoolAttr, "false">:$reassociate_fp_reductions,
-     DefaultValuedAttr<BoolAttr, "false">:$enable_index_optimizations,
-     DefaultValuedAttr<BoolAttr, "false">:$enable_arm_neon,
-     DefaultValuedAttr<BoolAttr, "false">:$enable_arm_sve,
-     DefaultValuedAttr<BoolAttr, "false">:$enable_amx,
-     DefaultValuedAttr<BoolAttr, "false">:$enable_x86vector,
-     DefaultValuedAttr<BoolAttr, "false">:$enable_async);
-  let results = (outs TransformHandleTypeInterface:$transformed);
-  let assemblyFormat = "$target attr-dict `:` functional-type($target, results)";
-  let cppNamespace = "mlir::transform_ext";
-}
-
 def RegisterMatchCallbacksOp :
     Op<Transform_Dialect, "iree.register_match_callbacks",
       [DeclareOpInterfaceMethods<MemoryEffectsOpInterface>,
