+++ llvm-external-projects/iree-dialects/lib/Dialect/LinalgTransform/IR/CMakeLists.txt
@@ -1,6 +1,4 @@
 add_mlir_library(IREELinalgTransformDialect
-  LinalgTransformOps.cpp
-  ScopedTransform.cpp
   StructuredTransformOpsExt.cpp
 
   DEPENDS
