--- llvm-external-projects/iree-dialects/python/iree/compiler/dialects/LinalgTransformBinding.td
@@ -1,12 +1,0 @@
-// Copyright 2021 The IREE Authors
-//
-// Licensed under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#ifndef PYTHON_BINDINGS_IREE_LINALGTRANSFORM_BINDING
-#define PYTHON_BINDINGS_IREE_LINALGTRANSFORM_BINDING
-
-include "iree-dialects/Dialect/LinalgTransform/LinalgTransformOps.td"
-
-#endif // PYTHON_BINDINGS_IREE_LINALGTRANSFORM_BINDING
