+++ llvm-external-projects/iree-dialects/test/lib/CMakeLists.txt
@@ -1,2 +1,1 @@
-add_subdirectory(Dialect)
 add_subdirectory(Transforms)
