--- llvm-external-projects/iree-dialects/test/lib/Dialect/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(LinalgTransform)
