--- llvm-external-projects/iree-dialects/test/lib/Dialect/LinalgTransform/CMakeLists.txt
@@ -1,12 +1,0 @@
-add_mlir_library(IREELinalgTransformTestPasses
-  TestScopedTransform.cpp
-
-  EXCLUDE_FROM_LIBMLIR
-
-  DEPENDS
-  mlir-headers
-
-  LINK_LIBS PUBLIC
-  IREELinalgTransformDialectPasses
-  MLIRPass
-  )
